django-subscriptions/pyproject.toml
[tool.poetry]
name = "django-subscriptions"
version = "0.1.0"
description = ""
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = "^0.46.0"
Django = "^3.1.7"
uvicorn = {extras = ["standard"], version = "^0.13.4"}

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
